commit! (0909)
commit! (0909)

+++ thymybatis/src/main/java/com/ss/thymybatis/dto/Comment.java
@@ -11,7 +11,7 @@ public class Comment {
 	private int id;    			// 댓글번호
 	private String nickname; 	// 닉네임
 	private String body;    	// 내용
-	private long articeId;    	// 글번호
+	private int articeId;    	// 글번호
 
 	public void setArticleId(Long id) {
 	}
